+++ src/layouts/App/Header/HeaderNav/index.tsx
@@ -1,4 +1,4 @@
-import styles from './HeaderNav.module.scss';
+import styles from './index.module.scss';
 
 const HeaderNav = () => {
   return (
