+++ src/layouts/App/Header/LogoutModal/index.module.scss
